+++ pkg/organization/organization.go
@@ -2,6 +2,7 @@ package organization
 
 import (
 	"fmt"
+	"os"
 	"strings"
 
 	templateorg "github.com/giantswarm/kubectl-gs/v2/pkg/template/organization"
@@ -67,6 +68,15 @@ func (o *Org) Build() (*orgv1alpha1.Organization, error) {
 	// This check will allow us to re-use existing orgs too without accidentally deleting the org when done
 	orgCR.ObjectMeta.Annotations = map[string]string{
 		DeleteAnnotation: "true",
+	}
+
+	// If found, populate details about Tekton run as labels
+	orgCR.ObjectMeta.Labels = map[string]string{}
+	if os.Getenv("TEKTON_PIPELINE_RUN") != "" {
+		orgCR.ObjectMeta.Labels["cicd.giantswarm.io/pipelinerun"] = os.Getenv("TEKTON_PIPELINE_RUN")
+	}
+	if os.Getenv("TEKTON_TASK_RUN") != "" {
+		orgCR.ObjectMeta.Labels["cicd.giantswarm.io/taskrun"] = os.Getenv("TEKTON_TASK_RUN")
 	}
 
 	orgCR.Status.Namespace = o.GetNamespace()